Make up{} work with GMP (#2986)
This is a little fussy, but with GMP, the labels are `pod` and
`namespace`. I have a label mapping that seems to work for scraped
metrics, but `up{}` is a special Prometheus metric and it doesn't seem
to respect the label mapping (not clear why, opening an internal bug
now).

Instead of `kubernetes_pod_name`, just use the relevant role labels.
In general this is probably a better choice, but I don't want to make
a wide change.