Add build report automation
Adds build/report/report.go, which generates a dated report and an
index.html that redirects to it.

You can see it live at https://agones-build-reports.storage.googleapis.com/index.html

The formatting is terrible, apologies for the web1.0 vibe.

Adds a cloudbuild.yaml to run it. I opted to strip the build down to
just the gcloud build runner because I was having difficulty with
authentication with the normal build container - plus it's _way_
heavier to build.