Rework health check handling of InitialDelaySeconds
See https://github.com/googleforgames/agones/issues/2966#issuecomment-1492346086:

* We remove any knowledge in the SDK of InitialDelaySeconds
* We remove the runHealth goroutine from main and shift this
responsibility to the /gshealthz handler

Along the way:

*  I noted that the FailureThreshold doesn't need to be enforced on
both the kubelet and SDK side, so in the injected liveness probe, I
dropped that to 1. Previously we were waiting more probes than we
needed to. In practice this is not terribly relevant since the SDK
pushes it into Unhealthy.