Extend e2e queue timings
This PR reworks the e2e timeouts to allow for more time for a given
build to wait to run e2es, but tightens the e2e deadline slightly:

* Tighten the per-e2e-configuration testcase to 1.5h. e2es are coming
in close to an hour in some cases but now that we're not running
consul, we don't need it as high as 2h. I don't think it's worth
tightening this all the way to an hour, though it would probably work.
  * Also drops the queueTtl for the CI sub-builds, these should not be
  queued for long since we serialize e2es now.

* Extends the e2e-wait-to-become-leader timeout to 3h. In higher
traffic times, we're hitting this limit often now, which only results
in a vicious cycle of retrying PRs. Instead wait longer to become
leader.

* Bumps the global timeout to 5h after aggregating:
  3h (e2e-wait-to-become-leader) + 1.5h (e2e timeout) + 0.5h (everything else)

* Remove vestigates of consul - it's no longer running anywhere.