Have TestPlayerConnectWithCapacityZero use framework to wait
I'm seeing e2e failures in this test, which I'm trying to track
down, but it seems like when this test fails, it explodes, e.g.:

panic: Fail in goroutine after TestPlayerConnectWithCapacityZero has completed
[...]
 agones.dev/agones/test/e2e.TestPlayerConnectWithCapacityZero.func1()
     /go/src/agones.dev/agones/test/e2e/gameserver_test.go:1111 +0x2d3

I think this is largely due to using `require`, which panics. Instead,
we have a framework function that does exactly what this is looping
on, just use that.