sdkserver: When waitForConnection fails, container should restart quickly (#3157)
* sdkserver: When waitForConnection fails, container should restart quickly

* waitForConnection is using the context provided by
NewSDKServerContext, but that context isn't closed until we see a
Shutdown(). That's not right - we want to restart the SDK server
quickly, which might save the game server in this odd case. So, move
waitForConnection up into main so that we can use the "normal"
context.

* Additionally, I noticed that when we went into graceful termination,
the liveness probes stopped! Agghhh. So, this reverts part of #3072:
instead of using /gshealthz as our heartbeat, fall back to using a go
routine. But we still honor the intent of #3072: health checks should
start when we see the first /gshealthz.

* Along the way, add some logging, and make it consistently "SDK" in
logging.

---------