GKE Autopilot: Add primary container annotation for game server container (#3234)
GKE Autopilot: Add autopilot.gke.io/primary-container annotation for game server container

`autopilot.gke.io/primary-container` is a new annotation that will be
supported by Autopilot ~soon that indicates which container resource
defaulting should affect. Prior to this if I were to create a
GameServer with no resource requests, Autopilot would default-up the
sidecar container, as it occurs first.

The annotation will be supported in 1.27 ~soon, but this change is
safe on all versions (just not useful, necessarily).