Race Flake in TestControllerSyncFleetAutoscaler()
Weird that this has more consistently failed with the Go race condition
checker failing on this test recently -- but it's a test we wrote a
while ago!

The flake was that the test added a Reactor to the mock _after_ the
Informer system had been started - which caused a race condition inside
the Fake client system as different systems interacted with the mock.

Changing the order fixes the issue!