Fix for scaling split allocated GameServerSets
Fixes bug wherein if a set of Allocations occurred across two or
more GameServerSets that had yet to be deleted for a RollingUpdate
(because of Allocated GameServers), and a scale down operation moved
the Fleet replica count to below the current number of Allocated
GameServers -- scaling back up would not move above the current number
of Allocated GameServers.

Or to put it another way, the current Fleet update logic didn't consider
old GameServerSets with Allocated GameServers but a 0 value for
`Spec.Replicas` as a complete rollout when scaling back up, so the logic
went back into rolling update logic, and it all went sideways.

This short circuits that scenario up front.

Close #3287