APIService: Updates to handlers for 1.27.x
This includes a bug fix for the issue outlined in
https://github.com/kubernetes/kubernetes/issues/119662, specifically
returning HTTP 406 for root /apis for all requests.

Also included an empty OpenAPI v3 discovery document to `/openapi/v3`
to stop the regular requests from the K8s control plane, while also
bringing the implementation for `/openapi/v2` to be consistent with the
new one.

Snuck in a linter fix as well.

Closes #3172