Flaky: TestAllocatorAllocateOnGameServerUpdateError
No, but for-realsies, I think I got it this time.

So this is what I think the flake is:

On each retry in `Allocate` the attempted GameServer is removed from the
list of batched GameServers.

We can retry up to 5 times, but because Jitter and machine speed, often
it's less than that -- so the test passes with only 4 GamServers in the
list.

But if the unit test is feeling speedy, it can work through all the
potential `GameServers` and then come to the conclusion - nope, there
are none left, so I'll say I can't find one.

The answer: Provide more GameServers than the allocation retries limit.