Bugs and Improvements for CPP SDK and Example
This fixes a bug in the CPP SDK, that was highlighted in the CPP
example, but didn't end show up in the conformance tests, as well as
several quality of life improvements.

* Bring all the gRPC references to our current gRPC version in the
  project.
* Add `-j$(nproc)` for all `build` targets to utlise more cores when
compiling. Shaved some time off compilation.
* Make the setup and compilation of the example align with the unit
tests (where it makes sense) - so same base image, etc.
* Explicitly compile the Abseil dependency (this was the critical
issue!)

Not quite sure if there is a good way to force this kind of issue to
show up in the conformance tests, given that we rely on gRPC to be
installed on the base image all SDK images are derived from -- but at
least this gets us over the current hurdle.

Work on #3281