restapi generation: clean before generation so file reorganization doesn't leak files
Without this change, if `restapi` generates a file previously and then
you run generation again and the file is _not_ generated, the old file
will still be there. This then results in the `header` script adding
the boilerplate a second time, resulting a diff every time.

Along the way: the boilerplate is still 2022