Remove `stale` and `obsolete` from PR's on update (#3431)
Bug example: https://github.com/googleforgames/agones/issues/1782

Shouldn't have closed, but instead was. I assume because it was still
labeled as "obsolete", and it had been 30 days - but only the stale
label had been removed.

I added `labels-to-remove-when-unstale` to all the operations, since I'm
not sure which operation does the work (and figured it couldn't hurt).