Updates SDK server UpdateCounter to handle multiple changes in one request
as the REST API can request to change the capacity, count, and countdiff all in one request.

Also updates the SDK Server to follow the same change update pattern as the local SDK server of first updating the capacity, then setcount, then countdiff.

Fixes a bug in the local SDK server to properly update the setcount.