ci/cache project root cloudbuild.yaml fix
When setting up a trigger for this cloudbuild.yaml file, it can only run
from the repository root - which then means this script fails as files
aren't where they are expected.

This fix retains the original structure of the cloudbuild.yaml and
shifts the original source files into where the rest of the script
expects it.