Counts and Lists: Improvements to SDK docs
Much like the player tracking SDK documentation, I wanted to be explicit
in the SDK documentation for Counts and Lists where data was being
stored, and where default values where coming from with links back to
the generated API reference - so made some improvements to help
facilitate that.

This does still need a Counters and Lists landing page to reference
(this is my next task) as well, but wanted to get this done while I was
thinking about it.

Work on #2716