Nodepool upgrades on GKE Terraform apply
Per the docs:
"...node pools defined inside a cluster can't be changed (or
added/removed) after cluster creation without deleting and recreating
the entire cluster."

Which is not great - since you can end up with out-of-sync K8s versions
between the control plane and nodes, an inability to change nodepool
sizes and just a general lack of flexibility.

Moving the node pool definitions out of the cluster definition solves
this issue!

Closes #3339