Breaking: Remove Cmake gRPC install when not found
gRPC upgrades with gRPC where very brittle, and attempts to upgrade
dependencies and/or switch to a `FetchContent_Declare` type workflow
were becoming increasingly time-consuming - and not resulting with a
working solution.

Speaking with the gRPC team, they recommended not installing
dependencies within CMake (and how they do it admittedly an
antipattern). So this change removes the ability for the CMakeLists.txt
to install gRPC and makes finding the gRPC dependency a hard
requirement, and updates the documentation accordingly.

This change isn't onerous to the end user, as our SDK has only the one
dependency - gRPC (which does install all its dependencies). At some
point gRPC will have official support for vcpkg, at which point we
should probably move with them - but in the meantime, the cpp-simple
example has been updated with this change, and it's an extra ~4 lines.

This update also include a lot of other fixes:
* Upgrade from Debian Bullseye to Bookworm for all cpp related
containers
* Builds now actually happen in parallel, reducing build time to minutes
 not hours.
* Bump to gRPC 1.57.1
* Fix bug with gRPC dependency utf8_range.
* Move to C++17, because it's required for gRPC.

Work on #3570