Move GKEAutopilotExtendedDurationPods to Alpha in 1.28+
With 1.28+, we can comfortably support Agones on Autopilot. This
change moves `GKEAutopilotExtendedDurationPods` to Alpha, but
documents that it is supported in 1.28+ only. Additionally:

* I changed the e2e testing to support changing the gate only on
1.28+, including adding a step in the next version update
* I fixed alphabetization in feature stages

TODO: This gate will eventually require us to change documentation on
disruption controls, but that page is already so branching that I may
wait until the gate is beta at least.

Towards #3386