New recommendation for multi-cluster allocation
Offline we tend to recommend people use a Service Mesh for multi-cluster
allocation, but I keep seeing people running into our docs and trying to
implement our solution.

Rather than waiting on a full investigation and documentation, I wanted
to put a note at the top making a recommendation to go in a different
direction - so that people don't actively chase this path if they aren't
aware there are other (better) paths for multi-cluster communications.

Work on #2468