e2e: simplify the test for DirectToGameServer PodPolicy
Can't find a good way to send UDP traffic directly to a pod. This is not something that
k8s does out of the box and requires a product like Calico or similar that uses BGP to
provide direct access to pods. Removing the udp call and making the test much simpler