TestGameServerAllocationDuringMultipleAllocationClients: Readdress flake (#3831)
The actual bug is here:
  https://github.com/googleforgames/agones/blob/2023077aff4125b9d9822c5d20cd773d13138385/test/e2e/gameserverallocation_test.go#L1423

When we tried to accept errors in https://github.com/googleforgames/agones/pull/3750, we forgot to
change this t.Errorf to a simple log.

Along the way:

* change the timing of the scale-down so it happens more probably in
the middle of the allocations, which was the intent. By my guess, 3s
might be past all allocations. Also add a log to see if I'm right.

* minor cleanup

Towards #3553