agones-{extensions,allocator}: Be more defensive about draining (#3839)
I looked at the flakes for
TestGameServerCreationRightAfterDeletingOneExtensionsPod and
TestAllocatorAfterDeleteReplica, and looked at kubelet logs. We
definitely see the Pod go un-ready, but the timing is fairly tight and
I'm not sure how long it takes the Service to react. Let's be a little
more defensive here and lame-duck for a little longer.