agones-{extensions,allocator}: Pause after cancelling context (#3843)
* Revert "agones-{extensions,allocator}: Be more defensive about draining"

This reverts commit 68b04ee9dc59ca2d2f21118e8cadd754f9d2fc37.

* agones-{extensions,allocator}: Pause after cancelling context

After #3839 went in, we noticed the flakes in
TestAllocatorAfterDeleteReplica disappear, but
TestGameServerCreationRightAfterDeletingOneExtensionsPod remained. I
looked more closely at this and I developed a theory: My guess is that
`kube-apiserver` connections to webhooks are actually "sticky" using
http(s) keepalives. If the TCP connection never closes, we'd see the
behavior we do, which is the `EOF` from kube-apiserver going to write
on a dead socket.

I looked at how to close sockets on the server side to "finish" the
drain, but realized that we actually do, by cancelling the context.
The thing is that we cancel the context and immediately exit, but it
leaves no time for anything to shut down.

I tested this and it seems to drive away the flake, without adding in
the extra delay.

Reverts #3839