Flake: e2e/TestScheduleAutoscaler
Found that `nextCronMinuteBetween` would return an invalid crontab
of "59-0 * * * *" at minute 59 of the day. Implemented a fix so that
it now return "0-1 * * * *" which takes a little bit longer, but no
longer errors out.

Also updated test to fail fast rather than let the test continue if this
error happens, to aid in debugging in the future.