Controller for Pod in Succeeded state.
When sidecar containers are enabled, if a Pod `exit(0)`s, then it will
end up in a Succeeded state, and the backing GameServer will not get
deleted.

This introduces a new controller to check for this Pod state and move
the `GameServer` to a `Shutdown` state.

Also updates all records of simple-game-server to the latest version,
since we need it to test that this fix works.

Closes #4188