Replace bitname/kubectl with alpine/kubectl
Since Bitnami is doing stuff with their images, moving this to the
alpine/kubectl image instead.

It's a little ahead of what we use version wise -- but the helm
job is very basic, and this is updated regularly, has a sh shell
and has multiple architectures by default.

Replaces PR #4242 which seems to have stagnated.