cleanup(e2e): Scale back autoscaler timeout. (#4312)
Since a unit test in Go cannot pass 10m (there seems to be no setting
for this!) the timeout of 10m is not useful for Autopilot clusters
since it will cause a timeout panic if anything times out by 10m,
which also stops the e2e test runner from rerunning the tests.

So some minor cleanup on TestAllocatorAfterDeleteReplica since that
tends to fail the most, and updated timeout for Autopilot clusters.

Co-authored-by: Thomas Lacroix <thomas.lacroix@ubisoft.com>