feat(autoscaler)!: Remove caBundle requirement for HTTPS URLs
BREAKING CHANGE: HTTPS webhook and WASM URLs no longer require caBundle
when configured. The impact of this is that HTTPS URLs for a webhook
or wasm package that are served via a trusted certificate authority
will be processed successfully. This is particularly important for
Wasm based autoscaling, as WASM packages will likely be served from
Cloud storage providers, via HTTPS which is not controlled by the end
user.

The caBundle field is now only required for URLs using custom or
self-signed certificates, as they will fail when being called from
the Fleet Autoscaler.

This simplifies configuration for standard HTTPS endpoints
while maintaining security for custom certificate scenarios.

Changes:
- Updated URL validation logic to allow HTTPS without caBundle
- Modified HTTP client to be stored per Fleet Autoscaler,
  allowing for a custom http client per Fleet Autoscaler.
- Updated tests to verify new behavior
- Adjusted validation error messages