More fixes for SuperTuxKart example to attempt to fix flakiness. (#4359)
I've run this several hundred times on my Minikube cluster, and so far
it has failed to fail!

My theory is that the log file is doing something weird with the log
file before it really starts pushing logs to it -- and therefore
occasionally will assume there is no log at all.

To fix this I made the following changes:

- Bump version from 0.20 to 0.21 in Makefile, fleet.yaml, and
  gameserver.yaml
- Update Agones SDK from v1.51.0 to v1.53.0
- Replace deprecated hpcloud/tail with nxadm/tail v1.4.11
- Fix Dockerfile syntax: change 'as' to 'AS' for build stage
- Enhance log file monitoring with existence and size checks before
  tailing

Work on #4321

Co-authored-by: Thomas Lacroix <thomas.lacroix@ubisoft.com>